refactor(src/redux): changes in Rtoolkit
