+++ client/src/Components/Store/index.jsx
@@ -1,6 +1,6 @@
 import React, { useEffect } from "react";
 import { useDispatch, useSelector } from "react-redux";
-import { getProducts } from "../../redux/actions/productsAction";
+import { getProducts } from "../../redux/actions/productsActions";
 
 function Store() {
   const dispatch = useDispatch();
